requirements.txt
appdirs==1.4.4
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.5
Babel==2.9.1
backports.entry-points-selectable==1.1.0
black==20.8b1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
crispy-tailwind==0.2.0
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
dj-rest-auth
Django==3.2.8
django-allauth
django-cors-headers
django-crispy-forms==1.10.0
django-debug-toolbar
django-debug-toolbar-request-history==0.1.3
django-environ==0.4.5
django-extensions
django-heroku
django-filter
django-jazzmin
django-mathfilters==1.0.0
django-phonenumber-field
django-summernote
django-tailwind
django_s3_storage
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
djoser==2.0.5
Faker==6.5.2
filelock==3.0.12
gunicorn==20.1.0
httpie==2.5.0
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
oauthlib==3.1.1
packaging==21.0
pathspec==0.9.0
phonenumbers==8.12.35
Pillow==8.4.0
pipenv==2020.11.15
platformdirs==2.4.0
pluggy==0.13.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.3
PySocks==1.7.1
pytest==6.2.3
pytest-cov==2.11.1
pytest-django==4.2.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.4.1
regex==2021.8.21
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
shortuuid==1.0.1
six==1.16.0
source==1.2.0
sqlparse==0.4.2
sslcommerz-lib==1.0
text-unidecode==1.3
toml==0.10.2
twilio==7.2.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.7
validate-email==1.3
version==0.1.0
virtualenv==20.4.2
virtualenv-clone==0.5.4
waitress==2.0.0
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.11.2
django-storages
boto3
